🚚 [rename/#59] - Footer, Navigation.tsx 폴더 이동
