--- src/components/Footer.tsx
@@ -1,11 +1,0 @@
-function Footer() {
-    return (
-        <div className="mt-28 relative bottom-0 flex justify-center items-center gap-2 mt-12 pb-12 select-none">
-            <footer className="">
-                © 2024. 원장님9해주세요 😱 Co. all rights reserved.
-            </footer>
-        </div>
-    );
-}
-
-export default Footer;
