+++ src/components/layout/Footer.tsx
@@ -1,0 +1,9 @@
+function Footer() {
+  return (
+    <footer className="relative bottom-0 flex justify-center items-end mt-12 pb-6 select-none">
+      &copy; 2024. 원장님9해주세요 😱 Co. all rights reserved.
+    </footer>
+  );
+}
+
+export default Footer;
